Avoid use of GNOME when gnome-open isn't present
It's possible for some GNOME components to be in use and
GNOME_DESKTOP_SESSION_ID to be set even though gnome-open is not
installed.

Modern Ubuntu installations are like this (reproduced on Trusty).

The result is that Launchy will incorrectly determine the current
environment is GNOME-like, only to find "gnome-open" missing and then
fall back on "fallback_browsers" instead of using a reasonable
alternative like "Xdg", and that means the user's preferred default
browser (or whatever) won't be used.